refactor(integrations): merge client.go and server.go (#1067)
Only new integrations: Height, Linear, MS, MS Teams, Zoom.

Purely cosmetic, no functional change, but it's a small step in the path
to refactor integrations and their API, to have simpler templates.
