--- integrations/common/doc.go
@@ -1,2 +1,0 @@
-// Package common provides common utilities for integrations.
-package common
