+++ integrations/common/integration.go
@@ -1,3 +1,4 @@
+// Package common provides common utilities for integrations.
 package common
 
 import (
@@ -6,8 +7,23 @@ import (
 	"net/http"
 
 	"go.autokitteh.dev/autokitteh/internal/backend/muxes"
+	"go.autokitteh.dev/autokitteh/internal/kittehs"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
 )
+
+func Descriptor(uniqueName, displayName, logoURL string) sdktypes.Integration {
+	return kittehs.Must1(sdktypes.StrictIntegrationFromProto(&sdktypes.IntegrationPB{
+		IntegrationId: sdktypes.NewIntegrationIDFromName(uniqueName).String(),
+		UniqueName:    uniqueName,
+		DisplayName:   displayName,
+		LogoUrl:       logoURL,
+		ConnectionUrl: "/" + uniqueName,
+		ConnectionCapabilities: &sdktypes.ConnectionCapabilitiesPB{
+			RequiresConnectionInit: true,
+			SupportsConnectionTest: true,
+		},
+	}))
+}
 
 // ServeStaticUI registers an integration's static web content to
 // AutoKitteh's internal user-authenticated HTTP server. User auth
