+++ integrations/height/integration.go
@@ -1,28 +1,38 @@
-package microsoft
+package height
 
 import (
 	"go.uber.org/zap"
 
 	"go.autokitteh.dev/autokitteh/integrations/common"
-	"go.autokitteh.dev/autokitteh/integrations/microsoft/connection"
 	"go.autokitteh.dev/autokitteh/internal/backend/muxes"
+	"go.autokitteh.dev/autokitteh/sdk/sdkintegrations"
+	"go.autokitteh.dev/autokitteh/sdk/sdkmodule"
 	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
 	"go.autokitteh.dev/autokitteh/web/static"
 )
 
-// Start initializes all the HTTP handlers of all the Microsoft integrations. This
-// includes connection UIs, connection initialization webhooks, and event webhooks.
+var desc = common.Descriptor("height", "Height", "/static/images/height.png")
+
+// New defines an AutoKitteh integration, which
+// is registered when the AutoKitteh server starts.
+func New(v sdkservices.Vars) sdkservices.Integration {
+	return sdkintegrations.NewIntegration(
+		desc, sdkmodule.New(), status(v), test(v),
+		sdkintegrations.WithConnectionConfigFromVars(v))
+}
+
+// Start initializes all the HTTP handlers of the integration.
+// This includes an internal connection UI, webhooks for AutoKitteh
+// connection initialization, and asynchronous event webhooks.
 func Start(l *zap.Logger, m *muxes.Muxes, v sdkservices.Vars, o sdkservices.OAuth, d sdkservices.DispatchFunc) {
-	common.ServeStaticUI(m, desc, static.MicrosoftWebContent)
+	common.ServeStaticUI(m, desc, static.HeightWebContent)
 
 	h := newHTTPHandler(l, v, o, d)
 	common.RegisterSaveHandler(m, desc, h.handleSave)
 	common.RegisterOAuthHandler(m, desc, h.handleOAuth)
 
-	// Event webhooks (no AutoKitteh user authentication by definition, because
+	// TODO: Event webhooks (no AutoKitteh user authentication by definition, because
 	// these asynchronous requests are sent to us by third-party services).
-	m.NoAuth.HandleFunc("POST "+connection.ChangePath, h.handleEvent)
-	m.NoAuth.HandleFunc("POST "+connection.LifecyclePath, h.handleLifecycle)
 }
 
 // handler implements several HTTP webhooks to save authentication data, as
