--- integrations/linear/client.go
@@ -1,18 +1,0 @@
-package linear
-
-import (
-	"go.autokitteh.dev/autokitteh/integrations/common"
-	"go.autokitteh.dev/autokitteh/sdk/sdkintegrations"
-	"go.autokitteh.dev/autokitteh/sdk/sdkmodule"
-	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
-)
-
-var desc = common.Descriptor("linear", "Linear", "/static/images/linear.svg")
-
-// New defines an AutoKitteh integration, which
-// is registered when the AutoKitteh server starts.
-func New(v sdkservices.Vars) sdkservices.Integration {
-	return sdkintegrations.NewIntegration(
-		desc, sdkmodule.New(), status(v), test(v),
-		sdkintegrations.WithConnectionConfigFromVars(v))
-}
